# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: CortexAI POS API
  description: |
    API REST do CortexAI POS — sistema de Ponto de Venda com inteligência artificial.
    Gerencia produtos, vendas, clientes, cupons, categorias, estoque e processamento de texto com IA.
  version: 6.0.0
  contact:
    name: CortexAI
servers:
  - url: http://localhost:3001
    description: Servidor de desenvolvimento
paths:
...
  /api/produtos:
    get:
      summary: Listar produtos (com paginação)
      tags: [Produtos]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: q
          schema: { type: string }
          description: Busca por nome, SKU, código de barras ou categoria
        - in: query
          name: categoria
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Lista de produtos com paginação
  /api/produto:
    post:
      summary: Criar novo produto
      tags: [Produtos]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarProduto'
      responses:
        '200':
          description: Produto criado com sucesso
        '400':
          description: Dados inválidos
    put:
      summary: Atualizar produto
      tags: [Produtos]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarProduto'
      responses:
        '200':
          description: Produto atualizado
    delete:
      summary: Deletar produto
      tags: [Produtos]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome: { type: string }
      responses:
        '200':
          description: Produto deletado

  # ====== CATEGORIAS ======
...
  /api/notificacoes:
    get:
      summary: Listar notificações
      tags: [Notificações]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: tipo
          schema: { type: string }
        - in: query
          name: lida
          schema: { type: boolean }
      responses:
        '200':
          description: Notificações paginadas
  /api/notificacoes/lidas:
    put:
      summary: Marcar todas notificações como lidas
      tags: [Notificações]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Todas marcadas como lidas
  /api/notificacao/{id}/lida:
    put:
      summary: Marcar notificação como lida
      tags: [Notificações]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Notificação marcada como lida
  /api/notificacoes/nao-lidas/count:
    get:
      summary: Contar notificações não lidas
      tags: [Notificações]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Contagem de não lidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }

  # ====== RELATÓRIOS ======
  /api/relatorios/faturamento:
    get:
      summary: Relatório de faturamento por período
      tags: [Relatórios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: periodo
          schema: { type: string, enum: [semana, mes, ano], default: '30 dias' }
        - in: query
          name: inicio
          schema: { type: string, format: date }
        - in: query
          name: fim
          schema: { type: string, format: date }
      responses:
        '200':
          description: Dados de faturamento
  /api/relatorios/top-produtos:
    get:
      summary: Top produtos mais vendidos
      tags: [Relatórios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Ranking de produtos
  /api/relatorios/vendas-por-dia:
    get:
      summary: Vendas agrupadas por dia (range customizado)
      tags: [Relatórios]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: inicio
          required: true
          schema: { type: string, format: date }
        - in: query
          name: fim
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Vendas por dia
...
  /api/processar-texto:
    post:
      summary: Processar texto com IA (fuzzy match com produtos)
      tags: [IA]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [texto]
              properties:
                texto:
                  type: string
                  example: "2 Coca Cola, 1 agua mineral"
      responses:
        '200':
          description: Resultado do processamento com IA
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso: { type: boolean }
                  audioLogId: { type: integer }
                  texto: { type: string }
                  itens:
                    type: array
                    items:
                      type: object
                      properties:
                        textoOriginal: { type: string }
                        produto: { type: object, nullable: true }
                        quantidade: { type: integer }
                        confianca: { type: integer }
                        encontrado: { type: boolean }
                  resumo:
                    type: object
                    properties:
                      totalItens: { type: integer }
                      encontrados: { type: integer }
                      confiancaMedia: { type: integer }
                      totalEstimado: { type: number }
  /api/audio-logs:
    get:
      summary: Listar logs de processamento de IA
      tags: [IA]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Logs de áudio

  # ====== ESP32 ======
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CriarProduto:
      type: object
      required: [nome, preco]
      properties:
        nome: { type: string }
        estoque: { type: integer, default: 0 }
        preco: { type: number }
        categoria: { type: string }
        sku: { type: string }
        codigoBarras: { type: string }
        precoCusto: { type: number }
        estoqueMinimo: { type: integer }
        fornecedor: { type: string }
        validade: { type: string, format: date-time }
        lote: { type: string }
        localizacao: { type: string }
        status: { type: string, default: ativo }
        descricao: { type: string }
        tags: { type: string }
    AtualizarProduto:
      type: object
      required: [nome]
      properties:
        nome: { type: string }
        estoque: { type: integer }
        preco: { type: number }
        categoria: { type: string }
        sku: { type: string }
        status: { type: string }
...